feat: expand CI to do clippy and test

+++ crates/glowberry-lib/src/colored.rs
@@ -29,9 +29,9 @@ pub fn gradient(
 
     for &[r, g, b] in &*gradient.colors {
         colors.push(colorgrad::Color::from_linear_rgba(
-            f32::from(r),
-            f32::from(g),
-            f32::from(b),
+            r,
+            g,
+            b,
             1.0,
         ));
     }
@@ -79,7 +79,7 @@ pub fn gradient(
     for (x, y, pixel) in imgbuf.enumerate_pixels_mut() {
         let Color { r, g, b, .. } = grad.at(positioner(x, y) as f32);
 
-        *pixel = image::Rgb([r as f32, g as f32, b as f32]);
+        *pixel = image::Rgb([r, g, b]);
     }
 
     Ok(imgbuf)
